[fix](nereids)move validate data types before EliminateUnnecessaryProject rule
